% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welford.R
\docType{class}
\name{WelfordAlgorithm}
\alias{WelfordAlgorithm}
\title{Welford Algorithm for Incremental Mean, Variance e Standard Deviation}
\format{An \code{\link{R6Class}} object}
\description{
It handles a stream of numbers and computes descriptive values without a prior knowledge of how many values you have.
It basically computers values coming from a stream of numbers. 06 - January - 2020
}
\examples{
x <- WelfordAlgorithm$new()
x$push(4)
x$pushList(c(3,6.5,4))
x
y <- x$getDataValues()
y
print(x$numDataValues())
print(x$mean())
print(x$standardDeviation())
print(x$variance())

}
\author{
Alberto Calderone <sinnefa@gmail.com>
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WelfordAlgorithm$new()}}
\item \href{#method-clear}{\code{WelfordAlgorithm$clear()}}
\item \href{#method-push}{\code{WelfordAlgorithm$push()}}
\item \href{#method-pushList}{\code{WelfordAlgorithm$pushList()}}
\item \href{#method-getDataValues}{\code{WelfordAlgorithm$getDataValues()}}
\item \href{#method-numDataValues}{\code{WelfordAlgorithm$numDataValues()}}
\item \href{#method-mean}{\code{WelfordAlgorithm$mean()}}
\item \href{#method-variance}{\code{WelfordAlgorithm$variance()}}
\item \href{#method-standardDeviation}{\code{WelfordAlgorithm$standardDeviation()}}
\item \href{#method-print}{\code{WelfordAlgorithm$print()}}
\item \href{#method-clone}{\code{WelfordAlgorithm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Object initialization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$new(keep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keep}}{Keeps all values in memory. Default is FALSE to save memory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\subsection{Method \code{clear()}}{
Empties out the list of values and resets all calculations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\subsection{Method \code{push()}}{
Pushes one single element in the stream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$push(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value to push in the stream}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pushList"></a>}}
\subsection{Method \code{pushList()}}{
Pushes a list of values in the stream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$pushList(l)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{l}}{The list to push in the stream}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDataValues"></a>}}
\subsection{Method \code{getDataValues()}}{
Returns a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$getDataValues()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numDataValues"></a>}}
\subsection{Method \code{numDataValues()}}{
Return the number of values processed so far
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$numDataValues()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean"></a>}}
\subsection{Method \code{mean()}}{
Returns the mean
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$mean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-variance"></a>}}
\subsection{Method \code{variance()}}{
Return the variance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardDeviation"></a>}}
\subsection{Method \code{standardDeviation()}}{
Returns the standard deviation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$standardDeviation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints the current status of the stram
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordAlgorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
